requirements.txt
xmltodict==0.13.0
untangle==1.2.1
pandas==2.2.2
python-decouple==3.8
requests==2.31
openpyxl==3.1.2